useIssuedsQuery -> useTotalIssuedDebt - And just add types correct types manually to the fetch json resposonse
