--- v2/lib/useIssuedsQuery/index.ts
@@ -1,1 +1,0 @@
-export * from './useIssuedsQuery';
