+++ v2/lib/useTotalIssuedDebt/index.ts
@@ -1,0 +1,1 @@
+export * from './useTotalIssuedDebt';
